requirements.txt
fastapi 
starlette_exporter
sqlalchemy
pydantic==2.9.1
psycopg2
python-slugify
pyjwt
jinja2